% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{add_elements}
\alias{add_elements}
\title{Adds named elements to a list when they are not there already.}
\usage{
add_elements(input, ..., .eager = TRUE)
}
\arguments{
\item{input}{List. The input list to manipulate.}

\item{...}{Key value pairs to add to the list provided the key is not already
used.}

\item{.eager}{Logical; Should the \code{value}s be evaluated?}
}
\value{
A modified list.
}
\description{
Adds named elements to a list when they are not there already.
}
