% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{alist2}
\alias{alist2}
\title{Make lazy list from arguments.}
\usage{
alist2(...)
}
\arguments{
\item{...}{Parameters to put into the list.}
}
\value{
A lazy list.
}
\description{
Works with passed \code{...} parameters.
}
\keyword{internal}
