% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma.R
\name{allma}
\alias{allma}
\title{Run a meta-analysis with all corrections.}
\usage{
allma(yi, vi, data, alpha = c(0, 0.05, 1), ...)
}
\arguments{
\item{yi}{Numeric vector of length code{k} with observed effect size
estimates.}

\item{vi}{Numeric vector of length code{k} with sampling variances.}

\item{alpha}{Numeric vector; Specifies the cuttoffs for significance.
Should include 0 and 1. Defaults to (0, 0.05, 1).}

\item{...}{Passed to \code{rstan::sampling}.}
}
\description{
Do a publication bias-correct, a p-hacking-correct and a classical
   meta-analysis.
}
