% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{arguments}
\alias{arguments}
\title{Get arguments of the calling function.}
\usage{
arguments(expand_dots = FALSE)
}
\arguments{
\item{expand_dots}{Logical; If \code{TRUE}, returns the argument list with
expanded \code{...}.}
}
\value{
The unevaluated argument list given to the calling function.
}
\description{
Get arguments of the calling function.
}
\keyword{internal}
