% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma.R
\name{cma}
\alias{cma}
\title{Classical meta-analysis}
\usage{
cma(yi, vi, likelihood = c("normal", "fnormal"), data,
  effects = c("random", "fixed"), alpha = c(0, 0.025, 0.05, 1),
  prior = NULL, classical = FALSE, ...)
}
\arguments{
\item{yi}{Numeric vector of length code{k} with observed effect size
estimates.}

\item{vi}{Numeric vector of length code{k} with sampling variances.}

\item{likelihood}{String; Either a vector of length code{k} or a string
giving the likelihood for each observation. Only "normal" is supported.}

\item{data}{Optional list or data frame containing \code{yi}, \code{vi} and
\code{likelihood}.}

\item{effects}{The type of meta-analysis model to use. Valid choices are
"random" and "fixed". Currently only random effects are supported.}

\item{prior}{Optional list of prior parameters. See the details.}

\item{...}{Passed to \code{rstan::sampling}.}
}
\description{
Do a classical Bayesian meta-analysis. This is a wrapper for
    \code{ma}.
}
