% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat.cuddy2018}
\alias{dat.cuddy2018}
\title{Arabidopsis QTL data on gravitropism}
\format{An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.}
\source{
\href{https://phenome.jax.org/projects/Moore1b}{QTL Archive}
}
\usage{
data(dat.cuddy2018)
}
\description{
Data from a QTL experiment on gravitropism in
Arabidopsis, with data on 162 recombinant inbred lines (Ler x
Cvi). The outcome is the root tip angle (in degrees) at two-minute
increments over eight hours.
}
\examples{
data(grav)
times <- attr(grav, "time")
phe <- grav$pheno
\donttest{iplotCurves(phe, times)}
}
\references{
Moore et al. (2013) Genetics 195:1077-1086
(\href{https://www.ncbi.nlm.nih.gov/pubmed/23979570}{PubMed})
}
\keyword{datasets}
