% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-dang2018.R
\docType{data}
\name{dat.dang2018}
\alias{dat.dang2018}
\title{Dang 2018}
\format{The data frame contains the following columns:}
\source{
\href{https://phenome.jax.org/projects/Moore1b}{QTL Archive}
}
\usage{
data(dat.dang2018)
}
\description{
Results from 150 studies of ego depletion, the claim that self-control is a
limited resource which is tapped whenever self-control is exerted.
}
\references{
Moore et al. (2013) Genetics 195:1077-1086
(\href{https://www.ncbi.nlm.nih.gov/pubmed/23979570}{PubMed})
}
\keyword{datasets}
