% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{do_call}
\alias{do_call}
\title{Variant of do call with that preserves argument names.}
\usage{
do_call(.fn, .args = NULL, ..., .env = parent.frame())
}
\arguments{
\item{.fn}{Function to call.}

\item{.args}{List of arguments to \code{.fn}.}

\item{...}{Further arguments to \code{.fn}.}

\item{.env}{The environment where the call is to be evaluated.}
}
\value{
The effect of calling \code{.fn} with the supplied arguments in the
specified environment.
}
\description{
Variant of do call with that preserves argument names.
}
