% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{extract_parameter}
\alias{extract_parameter}
\alias{extract_theta0}
\alias{extract_theta}
\alias{extract_tau}
\alias{extract_eta}
\alias{extract_Isq}
\title{Extract parameters from an \code{ma} object.}
\usage{
extract_theta0(object, fun = mean)

extract_theta(object, fun = mean, i = NULL)

extract_tau(object, fun = mean)

extract_eta(object, fun = mean, i = NULL)

extract_Isq(object, fun = mean)
}
\arguments{
\item{object}{The fitted \code{ma} object.}

\item{fun}{The function to apply to the parameters.}

\item{i}{Optional index, specifying which parameter among many to apply fun
to. Only makes sense for \code{eta} and \code{theta}.}
}
\value{
The same kind of object as \code{fun} returns.
}
\description{
Extract parameters from an \code{ma} object.
}
