% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funnel.R
\name{forest}
\alias{forest}
\title{Forest Plot}
\usage{
forest(object, labeltext = NULL, ...)
}
\arguments{
\item{object}{The \code{mafit} object to plot.}

\item{labeltext}{Character vector describing each study. Defaults to
"Study 1", "Study 2", etc.}

\item{...}{Passed to \code{forestplot}.q1}
}
\value{
\code{NULL}, the function is called for its side effects.
}
\description{
Create a post-analaysis forest plot.
}
