% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funnel.R
\name{funnel}
\alias{funnel}
\alias{logfunnel}
\title{Funnel Plot with Significance Lines}
\usage{
funnel(object, ...)

logfunnel(yi, vi, data, likelihood, alpha, col = c("red", "blue"),
  pch = c(20, 20), ...)
}
\arguments{
\item{object}{The \code{mafit} object to plot.}

\item{...}{Passed to \code{plot}.}
}
\value{
\code{NULL}, the function is called for its side effects.
}
\description{
Create a funnel plot with lines for the significance thresholds.
}
