% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_estimation.R
\name{ll_psma}
\alias{ll_psma}
\title{Llikelihood function of publication bias and p-hacking models}
\usage{
ll_psma(yi, vi, theta0, tau, alpha, eta)
}
\arguments{
\item{yi}{Numeric vector of effect sies.}

\item{vi}{Numeric vector of variances.}

\item{theta0}{Numeric mean.}

\item{tau}{Numeric standard deviation.}

\item{alpha}{Numeric vector of thresholds.}

\item{eta}{Numeric vector of publication probabilities.}
}
\value{
Likelihood.
}
\description{
Llikelihood function of publication bias and p-hacking models
}
