% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\docType{methods}
\name{loo,mafit-method}
\alias{loo,mafit-method}
\title{Calculate the \code{loo} for an \code{ma} object.}
\usage{
\S4method{loo}{mafit}(x, ...)
}
\arguments{
\item{object}{The fitted \code{hma} object.}
}
\value{
A \code{loo} object.
}
\description{
Calculate the \code{loo} for an \code{ma} object.
}
