% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densities.R
\name{maps}
\alias{maps}
\alias{rmaps}
\alias{dmaps}
\title{Marginal Publication Selection Meta-analysis Model}
\usage{
rmaps(n, theta0, tau, sigma, alpha, eta)

dmaps(x, theta0, tau, sigma, alpha, eta, log = FALSE)
}
\arguments{
\item{n}{Number of observations. If \code{length(n) > 1}, the length is taken
to be the number required.}

\item{theta0}{Numeric vector; The mean of the underlying effect size
distribution.}

\item{tau}{Numeric vector; The standard deviation of the underlying effect
size distribution.}

\item{sigma}{Numeric vector; The standard deviation of the study, due to
sampling error.}

\item{alpha}{Numeric vector; Specifies the thresholds for publication
bias.}

\item{eta}{Numeric vector; Containing the probabilites of being a study
with the given p-value from being published. This is normalized so that
the maximumal element is 1.}

\item{x, q}{Numeric vector of quantiles.}

\item{log, log.p}{Logical; If \code{TRUE}, probabilities are given as
\code{log(p)}.}

\item{p}{Numeric vector of probabilities.}

\item{lower.tail}{Logical; If \code{TRUE}, the lower tail is returned.}
}
\description{
Density, distribution, quantile, random variate generation, and expectation
    calculation for the marginalized distribution for the publication
    selection meta-analysis model
}
\details{
The effect size distribution for the publication selection model is not
    normal, but has itself been selected for. These functions assume a
    normal underyling effect size distribution and one-sided selection on the
    effects.
}
