% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{ml_psma}
\alias{ml_psma}
\title{Maximum likelihood estimation of publication bias and p-hacking models}
\usage{
ml_psma(yi, vi, alpha = c(0, 0.025, 0.05, 1))
}
\arguments{
\item{yi}{Numeric vector of effect sies.}

\item{vi}{Numeric vector of variances.}

\item{alpha}{Numeric vector of thresholds.}
}
\value{
List of maximum likelihood estimates.
}
\description{
Maximum likelihood estimation of publication bias and p-hacking models
}
