% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma.R
\name{phma}
\alias{phma}
\title{Meta-analysis with p-hacking}
\usage{
phma(yi, vi, likelihood = c("normal", "fnormal"), data,
  effects = c("random", "fixed"), alpha = c(0, 0.05, 1),
  parameters = NULL, classical = FALSE, ...)
}
\arguments{
\item{yi}{Numeric vector of length code{k} with observed effect size
estimates.}

\item{vi}{Numeric vector of length code{k} with sampling variances.}

\item{likelihood}{String; Either a vector of length code{k} or a string
giving the likelihood for each observation. Valid choices are "normal"
and "fnormal".}

\item{data}{Optional list or data frame containing \code{yi}, \code{vi} and
\code{likelihood}.}

\item{effects}{The type of meta-analysis model to use. Valid choices are
"random" and "fixed". Currently only random effects are supported.}

\item{alpha}{Numeric vector; Specifies the cuttoffs for significance.
Should include 0 and 1. Defaults to (0, 0.05, 1).}

\item{parameters}{Optional list of prior parameters. See the details.}

\item{classical}{Logical; If \code{TRUE}, runs a classical meta-analysis. If
\code{FALSE}, runs a Hedges meta-analysis. Currently not supported.}

\item{...}{Passed to \code{rstan::sampling}.}
}
\description{
Do a meta-analysis that corrects for p-hacking. This is a wrapper for
    \code{ma}.
}
