% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densities.R
\name{phnorm}
\alias{phnorm}
\alias{dphnorm}
\alias{rphnorm}
\alias{pphnorm}
\title{p-hacking Meta-analysis Model}
\usage{
dphnorm(x, theta, sigma, alpha, eta, log = FALSE)

rphnorm(n, theta, sigma, alpha, eta)

pphnorm(q, theta, sigma, alpha, eta, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{x, q}{Numeric vector of quantiles.}

\item{theta}{Numeric vector; The mean of the underlying normal distribution.}

\item{sigma}{Numeric vector; The standard deviation of the study, due to
sampling error.}

\item{alpha}{Numeric vector; Specifies the thresholds for publication
bias.}

\item{eta}{Numeric vector; The mixing probability for each component.}

\item{log, log.p}{Logical; If \code{TRUE}, probabilities are given as
\code{log(p)}.}

\item{n}{Number of observations. If \code{length(n) > 1}, the length is taken
to be the number required.}

\item{lower.tail}{Logical; If \code{TRUE}, the lower tail is returned.}

\item{p}{Numeric vector of probabilities.}
}
\description{
Density, distribution, quantile, random variate generation, and expectation
    calculation for the distribution for the publication selection
    meta-analysis model
}
\details{
These functions assume one-sided selection on the effects, and do not take
    the effect size distribution into account. For that, see \code{mph}.
}
