% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\docType{methods}
\name{predict,mafit-method}
\alias{predict,mafit-method}
\title{Predict method for Publication Selection Meta-analysis Fits}
\usage{
\S4method{predict}{mafit}(object, newdata = NULL, type = c("yi",
  "effect"), observed = FALSE, fun, Nreps = 1000, ...)
}
\arguments{
\item{object}{Oject of class inheriting from \code{mafit}.}

\item{newdata}{Either a vector of variances or a function of no arguments
generating a fixed length vector of variances.}

\item{type}{The type of predicton. This could be the predicted \code{yi} or
the predicted effect.}

\item{observed}{Logical; If \code{TRUE}, the prediction is about the post-
selected effect or \code{yi}. If \code{FALSE}, the prediction is about
the pre-selected effect or \code{yi}}

\item{fun}{Optional function to apply on each column of the sample matrix.}

\item{Nreps}{Number of samples to draw from the posterior predictive
distribution.}
}
\value{
Matrix of samples from the posterior predictive distribution. The ith
    column contains samples from the ith vi.
}
\description{
Predict method for Publication Selection Meta-analysis Fits
}
