% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma.R
\name{psma}
\alias{psma}
\title{Meta-analysis with Publication Selection}
\usage{
psma(yi, vi, likelihood = c("normal", "fnormal"), data,
  effects = c("random", "fixed"), alpha = c(0, 0.025, 0.05, 1),
  prior = NULL, ...)
}
\arguments{
\item{yi}{Numeric vector of length code{k} with observed effect size
estimates.}

\item{vi}{Numeric vector of length code{k} with sampling variances.}

\item{likelihood}{String; Either a vector of length code{k} or a string
giving the likelihood for each observation. Only "normal" is supported.}

\item{data}{Optional list or data frame containing \code{yi}, \code{vi} and
\code{likelihood}.}

\item{effects}{The type of meta-analysis model to use. Valid choices are
"random" and "fixed". Currently only random effects are supported.}

\item{alpha}{Numeric vector; Specifies the cuttoffs for significance.
Should include 0 and 1. Defaults to (0, 0.025, 0.05, 1).}

\item{prior}{Optional list of prior parameters. See the details of \code{ma}.}

\item{...}{Passed to \code{rstan::sampling}.

donttest{
'model = psma(yi, vi, data = metafor::dat.begg1989,
prior = list(eta0 = c(3, 2, 1),
theta0_mean = 0.5,
theta0_sd = 10,
tau_mean = 1,
tau_sd = 1))
}}
}
\description{
Meta-analysis with Publication Selection
}
