% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densities.R
\name{psnorm}
\alias{psnorm}
\alias{dpsnorm}
\alias{ppsnorm}
\alias{rpsnorm}
\title{Publication Selection Meta-analysis Model}
\usage{
dpsnorm(x, theta, sigma, alpha, eta, log = FALSE)

ppsnorm(q, theta, sigma, alpha, eta, lower.tail = TRUE, log.p = FALSE)

rpsnorm(n, theta, sigma, alpha, eta)
}
\arguments{
\item{x, q}{Numeric vector of quantiles.}

\item{sigma}{Numeric vector; The standard deviation of the study, due to
sampling error.}

\item{alpha}{Numeric vector; Specifies the thresholds for publication
bias.}

\item{eta}{Numeric vector; Containing the probabilites of being a study
with the given p-value from being published. This is normalized so that
the maximumal element is 1.}

\item{log, log.p}{Logical; If \code{TRUE}, probabilities are given as
\code{log(p)}.}

\item{lower.tail}{Logical; If \code{TRUE}, the lower tail is returned.}

\item{n}{Number of observations. If \code{length(n) > 1}, the length is taken
to be the number required.}

\item{p}{Numeric vector of probabilities.}

\item{theta0}{Numeric vector; The mean of the underlying effect size
distribution.}

\item{tau}{Numeric vector; The standard deviation of the underlying effect
size distribution.}
}
\description{
Density, distribution, quantile, random variate generation, and expectation
    calculation for the distribution for the publication selection
    meta-analysis model
}
\details{
The effect size distribution for the publication selection model is not
    normal, but has itself been selected for. These functions assume
    one-sided selection on the effects. These functions do not assume the
    existence of an underlying effect size distribution. For these, see
    \code{maps}.
}
