% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publipha-package.R
\docType{package}
\name{publipha-package}
\alias{publipha-package}
\alias{publipha}
\title{The 'publipha' package.}
\description{
Meta-analysis that corrects for publication selection bias and
p-hacking.
}
\references{
Hedges, Larry V. "Modeling publication selection effects in meta-analysis."
Statistical Science (1992): 246-255.

Moss, Jonas and De Bin, Riccardo. "Modelling publication
bias and p-hacking" (2019) arXiv:1911.12445

Stan Development Team (2018). RStan: the R interface to Stan. R package
version 2.18.1. https://mc-stan.org
}
