% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{sd_ratio}
\alias{sd_ratio}
\title{Calculate the standard deviation ratio}
\usage{
sd_ratio(sigma, tau, theta0_sigma, vi)
}
\arguments{
\item{sigma}{Numeric; Standard deviation of the posterior mean for the
model.}

\item{tau}{Numeric vector; The standard deviaiton of the effect size
distribution. Set equal to zero for a fixed effects model. Is averaged
over if it is a vector.}

\item{theta0_sigma}{Numeric; Standard deviation for the prior on the mean
effect.}

\item{vi}{Numeric vector; Vector of study-specific variances.}
}
\value{
Numeric; The standard deviation ratio
}
\description{
Calculate the standard deviation for the posterior mean between two models.
   The first standard deviation is supplied, while the other is calculated
   as if it were from a random effects meta-analysis with a normal prior
   for the mean.
}
