% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma.R
\name{allma}
\alias{allma}
\title{Run a meta-analysis with all corrections.}
\usage{
allma(yi, vi, data, alpha = c(0, 0.025, 0.05, 1), ...)
}
\arguments{
\item{yi}{Numeric vector of length code{k} with observed effect size
estimates.}

\item{vi}{Numeric vector of length code{k} with sampling variances.}

\item{data}{Optional list or data frame containing \code{yi}, \code{vi} and
\code{likelihood}.}

\item{...}{Passed to \code{rstan::sampling}.}

\item{likelihood}{String; Either a vector of length code{k} or a string
giving the likelihood for each observation. Valid choices are "normal"
and "fnormal".}

\item{effects}{The type of meta-analysis model to use. Valid choices are
"random" and "fixed". Currently only random effects are supported.}

\item{prior}{Optional list of prior parameters. See the details.}

\item{classical}{Logical; If \code{TRUE}, runs a classical meta-analysis. If
\code{FALSE}, runs a Hedges meta-analysis. Currently not supported.}
}
\description{
Do a publication bias-correct, a p-hacking-correct and a classical
   meta-analysis.
}
