% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{loo,mafit-method}
\alias{loo,mafit-method}
\title{Calculate the \code{loo} for an \code{ma} object.}
\usage{
\S4method{loo}{mafit}(x, ...)
}
\arguments{
\item{x}{an object of class `mafit`.}

\item{...}{passed to [`loo`][loo::loo].}
}
\value{
A [`loo`][loo::loo] object.
}
\description{
Calculate the \code{loo} for an \code{ma} object.
}
