% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma.R
\name{ma}
\alias{ma}
\title{Meta-analysis Correcting for Publication Bias or p-hacking}
\usage{
ma(yi, vi, bias = c("publication selection", "p-hacking", "none"),
  likelihood = NULL, data, effects = c("random", "fixed"),
  alpha = c(0, 0.05, 1), parameters = NULL, classical = FALSE, ...)
}
\arguments{
\item{yi}{Numeric vector of length code{k} with observed effect size
estimates.}

\item{vi}{Numeric vector of length code{k} with sampling variances.}

\item{bias}{String; If "publication bias", corrects for publication bias. If
"p-hacking", corrects for p-hacking.}

\item{likelihood}{String; Either a vector of length code{k} or a string
giving the likelihood for each observation. Valid choices are "normal"
and "fnormal".}

\item{data}{Optional list or data frame containing \code{yi}, \code{vi} and
\code{likelihood}.}

\item{effects}{The type of meta-analysis model to use. Valid choices are
"random" and "fixed". Currently only random effects are supported.}

\item{alpha}{Numeric vector; Specifies the cuttoffs for significance.
Should include 0 and 1. Defaults to (0, 0.05, 1).}

\item{parameters}{Optional list of prior parameters. See the details.}

\item{classical}{Logical; If \code{TRUE}, runs a classical meta-analysis. If
\code{FALSE}, runs a Hedges meta-analysis. Currently not supported.}

\item{...}{Passed to \code{rstan::sampling}.}
}
\value{
An S4 object of class \code{mafit}.
}
\description{
Do a Bayesian meta-analysis. Correct for publication selection or p-hacking,
    or run an ordinary meta-analysis without correction.
}
