% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publipha-package.R
\docType{package}
\name{publipha-package}
\alias{publipha-package}
\alias{publipha}
\title{The 'publipha' package.}
\description{
Meta-analysis that corrects for publication selection bias and
    p-hacking.
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.18.1. http://mc-stan.org
}
