% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densities-snorm.R
\name{snorm}
\alias{snorm}
\alias{dsnorm}
\alias{rsnorm}
\alias{esnorm}
\title{Selected Normal Effect Size Distribution}
\usage{
dsnorm(x, theta0, tau, sigma, alpha = c(0, 0.025, 0.05, 1), eta,
  log = FALSE)

rsnorm(n, theta0, tau, sigma, alpha = c(0, 0.025, 0.05, 1), eta)

esnorm(theta0, tau, sigma, alpha, eta)
}
\arguments{
\item{x}{Numeric vector of quantiles.}

\item{theta0}{Numeric vector; The mean of the underlying effect size
distribution.}

\item{tau}{Numeric vector; The standard deviation of the underlying effect
size distribution.}

\item{sigma}{Numeric vector; The standard deviation of the study, due to
sampling error.}

\item{alpha}{Numeric vector; Specifies the thresholds for publication
bias.}

\item{eta}{Numeric vector; Containing the probabilites of being a study
with the given p-value from being published. This is normalized so that
the maximumal element is 1.}

\item{log}{Logical; If \code{TRUE}, probabilities are given as
\code{log(p)}.}

\item{n}{Number of observations. If \code{length(n) > 1}, the length is taken
to be the number required.}
}
\description{
Density, distribution, quantile, random variate generation, and expectation
    calculation for the effect size distribution for the selected effect size
    distribution of the one-sided normal discrete probability vector
    publication bias model.
}
\details{
The effect size distribution for the publication selection model is not
    normal, but has itself been selected for. These functions assume a
    normal underyling effect size distribution and one-sided selection on the
    effects.
}
